remove fairseq cli and modules from examples (#234)
* remove fairseq cli and modules from examples

* modify all cli import paths

* add cli for lightseq-deepspeed, delete useless __init__.py

* fix bug for test_ls_ops (bert encoder layer)

* format files

* remove examples from build

* remove examples cpp from cmakelist

* update version to 2.2.1

* Update CMakeLists.txt

recover cpp example build

+++ tests/fairseq_layers.py
@@ -606,7 +606,7 @@ def generate_bert_enc_layer():
         activation_dropout_ratio,
         activation_fn="gelu",
     )
-    layer.to(torch.device("cuda:0"))
+    layer.to(torch.device("cuda:0"), dtype=torch.half)
     return layer
